Fix some findbugs results.  One of which actually *did* turn out to be a bug.

+++ src/org/jruby/ext/ffi/StructLayout.java
@@ -457,7 +457,7 @@ public final class StructLayout extends Type {
         }
 
         private final long getOffset(int index) {
-            return offset + (index * arrayType.getComponentType().getNativeSize());
+            return offset + (long) (index * arrayType.getComponentType().getNativeSize());
         }
 
         private IRubyObject get(Ruby runtime, int index) {
